refactor(FE): 스타일파일 합치기
