+++ frontend/index.html
@@ -3,15 +3,7 @@
     <head>
         <meta charset="UTF-8" />
         <link rel="icon" type="image/svg+xml" href="/src/assets/vite.svg" />
-        <link rel="stylesheet" href="src/styles/common.css" />
-        <link rel="stylesheet" href="src/styles/home.css" />
-        <link rel="stylesheet" href="src/styles/joinin.css" />
-        <link rel="stylesheet" href="src/styles/login.css" />
-        <link rel="stylesheet" href="src/styles/reset.css" />
         <link rel="stylesheet" href="src/styles/style.css" />
-        <link rel="stylesheet" href="src/styles/main.css" />
-        <link rel="stylesheet" href="src/styles/attend.css" />
-        <link rel="stylesheet" href="src/styles/rental.scss" />
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
