+++ frontend/src/styles/attend.scss
