+++ frontend/src/styles/home.scss
