+++ frontend/src/styles/joinin.scss
