+++ frontend/src/styles/main.scss
